requirements.txt
requests
click
pyyaml
beautifulsoup4
black
tzdata
GitPython
PyGithub